kafka_services/producer/requirements.txt
# kafka_services/producer/requirements.txt
kafka-python==2.0.2
schedule==1.1.0
python-dotenv==0.19.2
requests==2.27.1
pandas==1.4.1
numpy==1.22.3
aiohttp==3.8.1
websockets==10.3
urllib3==1.26.7
python-dateutil==2.8.2
retry==0.9.2
kafka-python==2.0.2
python-binance>=1.0.0
pycoingecko>=3.0.0
python-dotenv>=0.19.0
schedule>=1.1.0
backoff>=2.1.0
prometheus-client>=0.14.0